Sync with latest

+++ basex-core/src/main/java/org/basex/query/up/expr/Transform.java
@@ -61,7 +61,7 @@ public final class Transform extends Arr {
 
   @Override
   public Value value(final QueryContext qc) throws QueryException {
-    final int o = (int) qc.resources.output.size();
+    final int o = qc.resources.output.size();
     final Updates updates = qc.resources.updates();
     final ContextModifier tmp = updates.mod;
     final TransformModifier pu = new TransformModifier();
